Melhorias na interface e correções de usabilidade
- Adicionado texto "Novo email" no rótulo de `Email.cshtml`.
- Estilização do formulário em `Location.cshtml` para responsividade.
- Mensagens de erro traduzidas para o português em `Location.cshtml.cs`.
- Aprimorada lógica de exclusão no `HamburgueresController.cs` para manter integridade referencial.
- Atualização da quantidade de bebidas no `PedidosController.cs`.
- Alterado tipo de dado de `Quantidade` em `Bebidas.cs` e `DBTecBurguerContext.cs` para `int`.
- Mensagem "FORA DE ESTOQUE" alterada para "ESGOTADO" em `Cardapio.cshtml`.
- Ajustes na lógica de adição ao carrinho em `Cart.cshtml`.
- Removido efeito de sombra do título "SUGESTÕES DO CHEFE" em `Index.cshtml`.
- Correção do item de menu "Cardapio" para "Cardápio" em `_Layout.cshtml`.
- Aprimoramentos no estilo do dropdown em `site.css`.
- Melhorias na lógica de exibição do dropdown de administração em `site.js`.
- Várias correções de estilo e formatação em `site.css`.

+++ TecBurguer/Models/Bebidas.cs
@@ -19,7 +19,7 @@ namespace TecBurguer.Models
 
         public string? Imagem { get; set; }
 
-        public decimal? Quantidade { get; set; }
+        public int? Quantidade { get; set; }
 
         public virtual ICollection<PedidoBebidas> PedidoBebida { get; set; }
     }
